requirements-dev.txt
appdirs==1.4.3
argcomplete==1.10.0
atomicwrites==1.3.0
attrs==19.1.0
backports.csv==1.0.7
beautifulsoup4==4.8.0
boto3==1.9.78
botocore==1.12.78
bs4==0.0.1
cached-property==1.5.1
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
deprecation==2.0.6
diff-match-patch==20181111
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2.4
Django-Admin-Object-Actions==0.1.6
django-allauth==0.40.0
django-bootstrap-form==3.4
django-crum==0.7.4
django-debug-toolbar==2.2
django-easy-pdf==0.1.1
django-filter==2.2.0
django-import-export==1.2.0
django-nested-admin==3.2.3
django-object-actions==1.0.0
django-outbox-base-layout==1.2
django-storages==1.7.1
djangorestframework==3.10.3
docutils==0.14
docx2txt==0.8
EbookLib==0.17.1
et-xmlfile==1.0.1
extract-msg==0.23.1
html5lib==1.0.1
idna==2.8
IMAPClient==2.1.0
importlib-metadata==0.19
isodate==0.6.0
jdcal==1.4.1
jmespath==0.9.3
lxml==4.4.0
more-itertools==7.2.0
nltk==3.4.5
oauthlib==3.1.0
odfpy==1.4.0
olefile==0.46
openpyxl==2.6.2
packaging==19.1
Pillow==6.1.0
pluggy==0.12.0
ply==3.11
py==1.8.0
pycep-correios==3.1.0
pycryptodome==3.8.2
PyJWT==1.7.1
pyparsing==2.4.2
PyPDF2==1.26.0
pytest==5.1.1
python-dateutil==2.7.5
python-decouple==3.1
python-monkey-business==1.0.0
python-pptx==0.6.18
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2019.2
PyYAML==5.1.2
reportlab==3.5.9
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
s3transfer==0.1.13
six==1.12.0
slate3k==0.5.3
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sortedcontainers==2.1.0
soupsieve==1.9.3
SpeechRecognition==3.8.1
sqlparse==0.3.0
static3==0.7.0
tablib==0.13.0
textract==1.6.3
tzlocal==1.5.1
unicodecsv==0.14.1
#urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==4.1
xhtml2pdf==0.2.3
xlrd==1.2.0
XlsxWriter==1.2.0
xlwt==1.3.0
zeep==3.4.0
#zipp==0.6.0
